requirements.txt
Django==3.1
django-extensions==1.6.1
dj-database-url==0.5.0
gunicorn==20.0.4
python-decouple==3.3
djangorestframework==3.11.0
model-bakery==1.1.1
pytest-django==3.9.0
drf-yasg==1.17.1
pytest-cov==2.10.0
psycopg2-binary==2.8.5
whitenoise==5.0.1
requests==2.24.0
